Move HTMLRenderer to text tree
